[!] Fix linux build cast uint64 to uint32 warning as error

+++ debugger/src/libdbg64g/services/exec/cmd/cmd_loadelf.cpp
@@ -48,7 +48,7 @@ void CmdLoadElf::exec(AttributeType *args, AttributeType *res) {
 
     DsuMapType *dsu = info_->getpDsu();
     uint64_t soft_reset = 1;
-    uint64_t addr = reinterpret_cast<uint32_t>(&dsu->ulocal.v.soft_reset);
+    uint64_t addr = reinterpret_cast<uint64_t>(&dsu->ulocal.v.soft_reset);
 
     tap_->write(addr, 8, reinterpret_cast<uint8_t *>(&soft_reset));
